+++ sdk/keyvault/azure_security_keyvault_secrets/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Bearer token authentication now requires TLS (https).
+
 ### Bugs Fixed
 
 ### Other Changes